lambda/provision/requirements.txt
pinecone-client
